Add basket service with Redis and MediatR implementation
This commit introduces the basket service which is an important microservice in the e-commerce domain. In this addition, the basket service utilizes Redis cache database for storing customer baskets. MediatR, AutoMapper, and StackExchange Redis are used in the new basket service for handling commands and queries, object-to-object mapping, and communication with Redis respectively. Additionally, related configurations are added to docker-compose files. The solution file has also been updated to include the new projects.