Add Discount service with CRUD operations
In this commit, a new "Discount" microservice has been added to the solution. The service involves the creation of Discount.API, Discount.Application, Discount.Core, and Discount.Infrastructure projects. The service includes the implementation of CRUD operations for discount coupons, related database configurations, and a Dockerfile for the service. In addition, the docker-compose files have been updated.